requirements.txt
streamlit
tensorflow
tf_keras
pillow
numpy